code/labs/ultimate_moe_inference/requirements.txt
# Ultimate MoE Inference Lab Dependencies

# Core ML
torch>=2.4.0
transformers>=4.45.0
accelerate>=0.34.0

# NVIDIA Optimizations
transformer-engine>=1.8
triton>=3.0.0

# Profiling & Monitoring
nvidia-ml-py>=12.0.0  # pynvml
pynvml>=11.5.0

# Data & Evaluation
datasets>=2.20.0
sentencepiece>=0.2.0
tiktoken>=0.7.0

# Utilities
pyyaml>=6.0
numpy>=1.26.0
pandas>=2.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Async support
aiohttp>=3.9.0

# Optional: vLLM comparison
# vllm>=0.6.0

